Update: int -> bigint

+++ dbmodel/product.go
@@ -25,8 +25,8 @@ type Product struct {
 	Id            int64     `rnsql:"id"  rntype:"INT"  rnopt:"PK NN UQ AI"  json:"id"`
 	BrandId       int64     `rnsql:"brand_id"  rntype:"INT"  rnopt:"NN"  FK:"BRAND.id"  json:"brand_id"`
 	Name          string    `rnsql:"name"  rntype:"VARCHAR(200)"  rnopt:"NN"  json:"name"`
-	Price         int64     `rnsql:"price"  rntype:"INT"  rnopt:"NN"  json:"price"`
-	Amount        int64     `rnsql:"amount"  rntype:"INT"  rnopt:"NN"  json:"amount"`
+	Price         int64     `rnsql:"price"  rntype:"BIGINT"  rnopt:"NN"  json:"price"`
+	Amount        int64     `rnsql:"amount"  rntype:"BIGINT"  rnopt:"NN"  json:"amount"`
 	TitleImageS3  string    `rnsql:"title_image_s3"  rntype:"VARCHAR(200)"  rnopt:"NN"  json:"title_image_s3"`
 	DescriptionS3 string    `rnsql:"description_s3"  rntype:"VARCHAR(200)"  rnopt:"NN"  json:"description_s3"`
 	CreatedTime   time.Time `rnsql:"created_time"  rntype:"DATETIME"  rnopt:"NN"  json:"created_time"`
